Extract tests for meta
These are now found tests/test_models_meta.py.
Also fixes for the class Meta as required for the tests
to pass.
