+++ tests/test_models.py
@@ -53,27 +53,6 @@ def test_allows_python_version_and_full():
     assert r.python_full_version == "8.1.9"
 
 
-HASH = "9aaf3dbaf8c4df3accd4606eb2275d3b91c9db41be4fd5a97ecc95d79a12cfe6"
-
-
-def test_meta():
-    m = models.Meta(
-        {
-            "hash": {"sha256": HASH},
-            "pipfile-spec": 6,
-            "requires": {},
-            "sources": [
-                {
-                    "name": "pypi",
-                    "url": "https://pypi.org/simple",
-                    "verify_ssl": True,
-                },
-            ],
-        }
-    )
-    assert m.hash.name == "sha256"
-
-
 @pytest.fixture()
 def sources():
     return models.SourceCollection(
